Implementing some structural changes with Jack.
