+++ chica/__init__.py
@@ -1,6 +1,6 @@
 
 from .utility import get_example_data_path
-from .coolant_geometry import coolant_geometry
-from .non_dimensional import non_dimensional
-from .runner import Solver
-from .setup import Setup
+from .coolant_geometry import toroidal_flow, poloidal_flow
+from .non_dimensional import film_coeff, nusselt
+from .runner import initial, looper
+from .geometry_setup import initial_setup, looper_setup
