+++ chica/utility.py
@@ -4,7 +4,8 @@ from os import path
 
 def get_example_data_path(relative_data_path):
 
-    data_path = path.join(path.dirname(__file__), relative_data_path)
+    data_path = path.join(path.dirname(__file__), 'example_data')
+    data_path = path.join(data_path, relative_data_path)
 
     if not path.isfile(data_path):
         raise FileNotFoundError("The following example data file could not be located, "
